Added files via upload
Variables in the data set are listed in order in the codebook. Variable names are identified with boldface, followed by the question wording or a short description of the variable, and coding of the values in italics below. Please note that a series of questions (identified with * before the variable name) were asked of half the sample only. The goal of this hackathon will be to generate a prediction for the presidential election. The dependent variable for this predictive model is presapproval. Please contact me at [kdonovan at sjfc dot edu] if you have any questions, and see you Thursday!